finish converting everything over
