+++ src/main/java/io/papermc/codebook/lvt/suggestion/StringSuggester.java
@@ -1,0 +1,71 @@
+/*
+ * codebook is a remapper utility for the PaperMC project.
+ *
+ * Copyright (c) 2023 Kyle Wood (DenWav)
+ *                    Contributors
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation;
+ * version 3 only, no later versions.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+ * USA
+ */
+
+package io.papermc.codebook.lvt.suggestion;
+
+import io.papermc.codebook.lvt.suggestion.context.ContainerContext;
+import io.papermc.codebook.lvt.suggestion.context.method.MethodCallContext;
+import io.papermc.codebook.lvt.suggestion.context.method.MethodInsnContext;
+import java.io.IOException;
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+public class StringSuggester implements LvtSuggester {
+
+    @Override
+    public @Nullable String suggestFromMethod(
+            final MethodCallContext call, final MethodInsnContext insn, final ContainerContext container)
+            throws IOException {
+        final String methodName = call.data().name();
+
+        if (methodName.startsWith("split")) {
+            if (insn.owner().name().equals("java/lang/String")
+                    || insn.owner().name().equals("com/google/common/base/Splitter")) {
+                return "parts";
+            }
+        }
+
+        if (methodName.equals("repeat")
+                && call.data().returnType().asInternalName().equals("Ljava/lang/String;")) {
+            return "repeated";
+        }
+        if (methodName.equals("indexOf") || methodName.equals("lastIndexOf")) {
+            return "index";
+        }
+        if (methodName.equals("substring")) {
+            return "sub";
+        }
+        if (methodName.equals("codePointAt")) {
+            return "code";
+        }
+        if (methodName.equals("trim")) {
+            return "trimmed";
+        }
+        if (methodName.startsWith("strip")) {
+            return "stripped";
+        }
+        if (methodName.equals("formatted")) {
+            return "formatted";
+        }
+
+        return null;
+    }
+}
