+++ src/main/java/io/papermc/codebook/lvt/suggestion/VerbPrefixBooleanSuggester.java
@@ -22,23 +22,40 @@
 
 package io.papermc.codebook.lvt.suggestion;
 
+import static io.papermc.codebook.lvt.suggestion.SuggestionUtil.tryMatchPrefix;
+
+import dev.denwav.hypo.model.data.types.PrimitiveType;
 import io.papermc.codebook.lvt.suggestion.context.ContainerContext;
 import io.papermc.codebook.lvt.suggestion.context.method.MethodCallContext;
 import io.papermc.codebook.lvt.suggestion.context.method.MethodInsnContext;
+import java.util.List;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
-public class GenericSuggester implements LvtSuggester {
+/*
+This matches against methods that return booleans with prefixes that suggest the full method name,
+including the prefix should be used as the local variable name.
+ */
+public class VerbPrefixBooleanSuggester implements LvtSuggester {
+
+    private static final List<String> BOOL_METHOD_PREFIXES = List.of("is", "has", "can", "should");
 
     @Override
     public @Nullable String suggestFromMethod(
             final MethodCallContext call, final MethodInsnContext insn, final ContainerContext container) {
-        return switch (call.data().name()) {
-            case "hashCode" -> "hashCode";
-            case "size" -> "size";
-            case "length" -> "len";
-            case "freeze" -> "frozen";
-            case "readLine" -> "line";
-            default -> null;
-        };
+        if (call.data().returnType() != PrimitiveType.BOOLEAN) {
+            return null;
+        }
+        final String methodName = call.data().name();
+
+        final @Nullable String prefix = tryMatchPrefix(methodName, BOOL_METHOD_PREFIXES);
+        if (prefix == null) {
+            return null;
+        }
+
+        if (Character.isUpperCase(methodName.charAt(prefix.length()))) {
+            return methodName;
+        } else {
+            return null;
+        }
     }
 }
